+++ CHANGELOG.md
@@ -7,6 +7,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+2.0.0 (2018-03-20)
+------------------
+### [Added]
+*   Add youtube Push Video
+
+### [Changed]
+*   Youtube::getVideos() function now use out Pull Api instead of the identities endpoints
+
+
 1.1.3 (2018-02-23)
 ------------------
 ### [Added]